requirements.txt
python-dotenv>=1.0.0,<2
requests>=2.31.0,<3
pandas>=2.0.0,<3
